added auth with jwt

+++ client/src/components/WorkoutDetails.js
@@ -1,14 +1,23 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 
-import formatDistanceToNow from 'date-fns/formatDistanceToNow' ; 
-
+import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+
+        if(!user){
+            return
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
@@ -22,8 +31,8 @@ const WorkoutDetails = ({ workout }) => {
             <h4>{workout.title}</h4>
             <p><strong>Load (kg): </strong>{workout.load}</p>
             <p><strong>Number of reps: </strong>{workout.reps}</p>
-            <p>{formatDistanceToNow(new Date(workout.createdAt) , {addSuffix : true})}</p>
-            <span className = "material-symbols-outlined" onClick={handleClick}>delete</span>
+            <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+            <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
         </div>
     )
 }
